+++ docs/redux.md
@@ -3,9 +3,9 @@
 Добавление данных:
 
 1. Создаем `actions` и `reducers` через `slice` в `@/utils/store/reducers`
-2. Тип данных добавляем интерфейсом в IStateScheme `@/utils/store/typings.ts`
+2. Тип новых данных добавляем интерфейсом в IStateScheme `@/typings/IStateScha.ts`
 3. Добавляем редюсер в `rootReducer` в `@/utils/store/store.ts`
-4. Описываем получаемые данные в `@/utils/store/selectors`
+4. Описываем данные, которые будем доставать из `state` в `@/utils/store/selectors`
 
 Использование в компанентах:
