+++ packages/client/src/typings/IAppStatus.ts
@@ -1,0 +1,4 @@
+export interface IAppStatus {
+  isLoading: boolean;
+  isAuth: boolean;
+}
