+++ packages/client/src/typings/IStateSchema.ts
@@ -1,0 +1,7 @@
+import { IAppStatus } from './IAppStatus';
+import { IUser } from './IUser';
+
+export interface IStateScheme {
+  user: IUser;
+  appStatus: IAppStatus;
+}
