+++ packages/client/src/typings/IUser.ts
@@ -1,0 +1,5 @@
+import { UserProfileForm } from '@/pages/userInfoPage/typings';
+
+export interface IUser extends UserProfileForm {
+  id: number;
+}
