+++ packages/client/src/utils/store/reducers/userSlice/userSlice.test.ts
@@ -1,5 +1,5 @@
 import { DeepPartial } from '@reduxjs/toolkit';
-import { IUser } from '../../typings';
+import { IUser } from '@/typings/IUser';
 import { userReducer, userActions } from './userSlice';
 
 const mockUser = {
