+++ packages/client/src/utils/store/selectors/getAppStatusSelectors/getAppStatusSelectors.test.ts
@@ -1,5 +1,5 @@
 import { DeepPartial } from '@reduxjs/toolkit';
-import { IAppStatus } from '../../typings';
+import { IAppStatus } from '@/typings/IAppStatus';
 import { getIsAuth, getIsLoading } from './getAppStatusSelectors';
 
 describe('getAppStatusSelectors', () => {
