+++ packages/client/src/utils/store/selectors/getUserFullSelector/getUserFullSelector.ts
@@ -1,3 +1,3 @@
-import type { IUser } from '../../typings';
+import type { IUser } from '@/typings/IUser';
 
 export const getUserFull = (state: IUser) => state;
