+++ packages/client/src/utils/store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore, ReducersMapObject } from '@reduxjs/toolkit';
 import { appStatusReducer } from './reducers/appStatusSlice/appStatusSlice';
 import { userReducer } from './reducers/userSlice/userSlice';
-import { IStateScheme } from './typings';
+import { IStateScheme } from '@/typings/IStateSchema';
 
 export function createReduxStore(initialState?: IStateScheme) {
   const rootReducer: ReducersMapObject<IStateScheme> = {
@@ -12,6 +12,6 @@ export function createReduxStore(initialState?: IStateScheme) {
   return configureStore<IStateScheme>({
     reducer: rootReducer,
     preloadedState: initialState,
-    // devTools: Добавить значение для отключения при продакт сборке
+    // devTools: Добавить значение `false` для отключения при продакт сборке
   });
 }
